feat: read token from config dir
Turns out relative paths in `read_to_string()` are relative from the
current directory when you execute the program, which isn't the root of
this project when you call the binary in another script.

So instead I'm making it read from a known location, namely under
`$XDG_CONFIG_HOME/lyrical`.