feat: add flake.nix
This required a change in how openssl is packaged for the `reqwest`
dependency that is used to make HTTP requests. Previously, I was using
"vendored" sources, meaning the openssl source code was fetched by
Cargo, and build using pkg-config and make on my host system (requiring
me to install make on my system, pkg-config apparently already was a
transient dependency from some other system package.). This approach was
causing issues with the Nix build system, as it wasn't set up to build
the provided openssl source. So instead of using the vendored version,
openssl is included as a runtime dependency explicitly.