fix: missing newline between some lyric text nodes
This fix makes sure that newline separation is consistent for all parts
of the song lyric's text.

An xpath expression is used to extract all the HTML elements that
contain song lyrics text at once, but genius.com song pages have a
non-lyrics segment in the middle of them. Up until now this caused the
following segment's text content to be appended to the output directly
so there would be no newline between some verses of the song.